tests/submission/PhilipTaylor/canvas/2d.strokeStyle.default.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: 2d.strokeStyle.default</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.strokeStyle.html">strokeStyle</a>.default</h1>
<p class="desc"></p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.2d.colours.default">2d.colours.default</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("");
_addTest(function(canvas, ctx) {

_assertEqual(ctx.strokeStyle, '#000000', "ctx.strokeStyle", "'#000000'");


});
</script>

<div id="log"></div>
</body></html>
